# backend/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "GET IT NOW API",
    "version": "1.0.0",
    "description": "On-demand service marketplace API for connecting customers with verified workers",
    "contact": {
      "name": "API Support",
      "email": "support@getitnow.example"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:4000",
      "description": "Development server"
    },
    {
      "url": "https://api.getitnow.example",
      "description": "Production server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/settlements": {
      "get": {
        "summary": "List settlements",
        "tags": [
          "Settlements"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "List of settlements"
          }
        }
      }
    },
    "/settlements/{id}/process": {
      "post": {
        "summary": "Process settlement (admin)",
        "tags": [
          "Settlements"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Settlement processed"
          },
          "404": {
            "description": "Not found"
          },
          "409": {
            "description": "Already processed"
          }
        }
      }
    },
    "/settlements/{id}": {
      "get": {
        "summary": "Get settlement details",
        "tags": [
          "Settlements"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Settlement details"
          },
          "404": {
            "description": "Not found"
          }
        }
      }
    },
...
    "/auth/password/set": {
      "post": {
        "summary": "Set initial password",
        "tags": [
          "Authentication"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Password set"
          }
        }
      }
    },
...
  }
}